+++ src/core.rs
@@ -2,6 +2,7 @@ pub mod channels;
 pub mod command_markers;
 pub mod common_conditions;
 pub mod ctx;
+pub mod event_registry;
 pub mod replication_registry;
 pub mod replication_rules;
 pub mod replicon_tick;
@@ -11,6 +12,7 @@ use serde::{Deserialize, Serialize};
 
 use channels::RepliconChannels;
 use command_markers::CommandMarkers;
+use event_registry::EventRegistry;
 use replication_registry::ReplicationRegistry;
 use replication_rules::ReplicationRules;
 
@@ -22,7 +24,8 @@ impl Plugin for RepliconCorePlugin {
             .init_resource::<RepliconChannels>()
             .init_resource::<ReplicationRegistry>()
             .init_resource::<ReplicationRules>()
-            .init_resource::<CommandMarkers>();
+            .init_resource::<CommandMarkers>()
+            .init_resource::<EventRegistry>();
     }
 }